products_api/v1/category/serializers.py
from rest_framework import serializers

from v1.category.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'
