requirements.txt
asgiref==3.4.1
Django==3.2.23
django-cors-headers==4.3.1
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
pip==18.1
psycopg2==2.9.5
pytz==2023.3.post1
setuptools==40.6.2
sqlparse==0.4.4
typing-extensions==4.1.1
